Dropped `toml` support and removed unnecessary abstract class
