+++ plugins/python/src/main/kotlin/io/paddle/plugin/python/PythonPlugin.kt
@@ -4,7 +4,6 @@ import io.paddle.plugin.Plugin
 import io.paddle.plugin.python.dependencies.authentication.AuthenticationProvider
 import io.paddle.plugin.python.dependencies.index.PyPackageRepositoryIndexer
 import io.paddle.plugin.python.extensions.*
-import io.paddle.plugin.python.tasks.generate.GeneratePyProject
 import io.paddle.plugin.python.tasks.generate.GenerateRequirements
 import io.paddle.plugin.python.tasks.install.CiTask
 import io.paddle.plugin.python.tasks.install.InstallTask
@@ -47,7 +46,6 @@ object PythonPlugin : Plugin {
             WheelTask(project),
             TwinePublishTask(project),
             GenerateRequirements(project),
-            GeneratePyProject(project),
         ) + RunTask.from(project) + PyTestTask.from(project)
     }
